+++ client/js/manage/controllers/manageCtrl.ng.js
@@ -75,4 +75,8 @@ angular.module('quick-survey').controller('ManageCtrl',
       });
     }
   };
+
+  $scope.share = function() {
+    window.parent.postMessage({'startSharing': {}}, '*');
+  };
 });