+++ client/js/manage/directives/question-details.ng.js
@@ -11,6 +11,9 @@ angular.module('quick-survey').directive('questionDetails', function () {
     },
     controller: function ($scope) {
       $scope.editingQuestion = false;
+
+      $scope.saved = false;
+
       $scope.editQuestion = function (question) {
         $scope.editingQuestion = !$scope.editingQuestion;
       };
@@ -20,7 +23,7 @@ angular.module('quick-survey').directive('questionDetails', function () {
       };
 
       $scope.deleteQuestion = function (index) {
-        var success = confirm("Are you sure you want to delete this question?")
+        var success = confirm("Are you sure you want to delete this question?");
         if (success) {
           $scope.survey.questions.splice(index, 1);
           $scope.save();
@@ -28,14 +31,20 @@ angular.module('quick-survey').directive('questionDetails', function () {
       };
 
       $scope.save = function() {
-        console.log($scope.survey.questions)
+        $scope.saved = false;
         Surveys.update($scope.survey._id,
           {$set: {
             questions: angular.copy($scope.survey.questions)
           }}, function(err, resp) {
-            if (err) console.log('error', err);
-            console.log('success', resp);
-          })
+            if (err) {
+              console.log('error', err);
+            } else {
+              $scope.$apply(function () {
+                $scope.saved = true;
+              });
+            }
+
+          });
       };
     },
     templateUrl: 'client/js/manage/directives/question-details.ng.html',